moved some functions to the gamma_support_functions.py file
